Ability to switch stream type, toolbar changes
