+++ src/prompts/prompts.js
@@ -1,8 +1,12 @@
 import filian from './filian'
 import neuro from './neuro'
 
+/**
+ * 1. The array name is the route path
+ * 2. Add the route to ../components/StreamTypeLinks.vue
+ */
 export default {
-  solo: [
+  justChatting: [
     neuro
   ],
